Split sequencer management
