+++ src/components/SequencerTracks.vue
@@ -6,9 +6,9 @@
 <script setup lang="ts">
 import { inject, computed } from 'vue';
 import Track from '@/components/Track.vue';
-import { SequencerManager } from '@/services/SequencerManager';
-import { sequencerManagerKey } from '@/utils/injection-keys';
+import { sequencerTrackManagerKey } from '@/utils/injection-keys';
+import { SequencerTrackManager } from '@/services/SequencerTrackManager';
 
-const sequencerManager = inject<SequencerManager>(sequencerManagerKey);
-const tracks = computed(() => sequencerManager?.tracks.value);
+const sequencerManager = inject<SequencerTrackManager>(sequencerTrackManagerKey) as SequencerTrackManager;
+const tracks = computed(() => sequencerManager.tracks.value);
 </script>
