+++ src/services/commands/SequencerCommands.ts
@@ -1,46 +1,48 @@
-export class AddTrackCommand {
+import { Command } from "@/utils/interfaces";
+
+export class AddTrackCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class RemoveTrackCommand {
+export class RemoveTrackCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class RemoveLastTrackCommand {
+export class RemoveLastTrackCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class ClearTracksCommand {
+export class ClearTracksCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class AddStepCommand {
+export class AddStepCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class RemoveStepCommand {
+export class RemoveStepCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
     redo() { }
 }
 
-export class RemoveLastStepCommand {
+export class RemoveLastStepCommand implements Command {
     constructor() { }
     execute() { }
     undo() { }
