+++ src/utils/injection-keys.ts
@@ -1,4 +1,8 @@
 import type { InjectionKey } from 'vue';
-import { SequencerManager } from '@/services/SequencerManager';
+import { SequencerInstrumentManager } from '@/services/SequencerInstrumentManager';
+import { SequencerPlaybackManager } from '@/services/SequencerPlaybackManager';
+import { SequencerTrackManager } from '@/services/SequencerTrackManager';
 
-export const sequencerManagerKey: InjectionKey<SequencerManager> = Symbol('sequencerManager');
+export const sequencerInstrumentManagerKey: InjectionKey<SequencerInstrumentManager> = Symbol('sequencerInstrumentManager');
+export const sequencerPlaybackManagerKey: InjectionKey<SequencerPlaybackManager> = Symbol('sequencerPlaybackManager');
+export const sequencerTrackManagerKey: InjectionKey<SequencerTrackManager> = Symbol('sequencerTrackManager');
